# api/templates/openapi.json
{
   "openapi": "3.1.0",
   "info": {
      "title": "biodata-registry-service",
      "description": "\n## biodata-registry-service\n\nService to fetch data from Registry.\n\n",
      "version": "0.0.0"
   },
   "paths": {
...
      "/specimen_procedure": {
         "post": {
            "tags": [
               "core"
            ],
            "summary": "Create Specimen Procedure",
            "operationId": "create_specimen_procedure",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpecimenProcedureCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/SpecimenProcedures"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Specimen Procedure",
            "operationId": "get_specimen_procedure",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/SpecimenProcedures"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "core"
            ],
            "summary": "Delete",
            "operationId": "delete_specimen_procedure",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "core"
            ],
            "summary": "Update",
            "operationId": "update_specimen_procedure",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpecimenProcedureUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/SpecimenProcedures"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/specimen_procedures": {
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Specimen Procedures",
            "operationId": "get_specimen_procedures",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/SpecimenProcedures"
                           },
                           "title": "Response Get Specimen Procedures"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/specimen": {
         "post": {
            "tags": [
               "core"
            ],
            "summary": "Create Specimen",
            "operationId": "create_specimen",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpecimenCreate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Specimens"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Specimen",
            "operationId": "get_specimen",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Specimens"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "core"
            ],
            "summary": "Delete",
            "operationId": "delete_specimen",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {}
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "core"
            ],
            "summary": "Update",
            "operationId": "update_specimen",
            "parameters": [
               {
                  "name": "id",
                  "in": "query",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "title": "Id"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/SpecimenUpdate"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Specimens"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/specimens": {
         "get": {
            "tags": [
               "core"
            ],
            "summary": "Get Specimens",
            "operationId": "get_specimens",
            "parameters": [
               {
                  "name": "offset",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "default": 0,
                     "title": "Offset"
                  }
               },
               {
                  "name": "limit",
                  "in": "query",
                  "required": false,
                  "schema": {
                     "type": "integer",
                     "maximum": 1000,
                     "default": 10,
                     "title": "Limit"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/Specimens"
                           },
                           "title": "Response Get Specimens"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "schemas": {
...
         "HTTPValidationError": {
            "properties": {
               "detail": {
                  "items": {
                     "$ref": "#/components/schemas/ValidationError"
                  },
                  "type": "array",
                  "title": "Detail"
               }
            },
            "type": "object",
            "title": "HTTPValidationError"
         },
...
         "SpecimenCreate": {
            "properties": {
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "subject_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Subject Id"
               }
            },
            "type": "object",
            "required": [
               "name"
            ],
            "title": "SpecimenCreate"
         },
         "SpecimenProcedureCreate": {
            "properties": {
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "title": "SpecimenProcedureCreate"
         },
         "SpecimenProcedureUpdate": {
            "properties": {
               "data": {
                  "anyOf": [
                     {
                        "additionalProperties": true,
                        "type": "object"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "title": "SpecimenProcedureUpdate"
         },
         "SpecimenProcedures": {
            "properties": {
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               }
            },
            "type": "object",
            "title": "SpecimenProcedures"
         },
         "SpecimenUpdate": {
            "properties": {
               "name": {
                  "anyOf": [
                     {
                        "type": "string",
                        "maxLength": 254
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Name"
               },
               "data": {
                  "anyOf": [
                     {
                        "additionalProperties": true,
                        "type": "object"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "subject_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Subject Id"
               }
            },
            "type": "object",
            "title": "SpecimenUpdate"
         },
         "Specimens": {
            "properties": {
               "id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Id"
               },
               "name": {
                  "type": "string",
                  "maxLength": 254,
                  "title": "Name"
               },
               "data": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Data"
               },
               "schema_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Schema Id"
               },
               "space_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Space Id"
               },
               "subject_id": {
                  "anyOf": [
                     {
                        "type": "integer"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Subject Id"
               }
            },
            "type": "object",
            "required": [
               "name"
            ],
            "title": "Specimens"
         },
...
         "ValidationError": {
            "properties": {
               "loc": {
                  "items": {
                     "anyOf": [
                        {
                           "type": "string"
                        },
                        {
                           "type": "integer"
                        }
                     ]
                  },
                  "type": "array",
                  "title": "Location"
               },
               "msg": {
                  "type": "string",
                  "title": "Message"
               },
               "type": {
                  "type": "string",
                  "title": "Error Type"
               },
               "input": {
                  "title": "Input"
               },
               "ctx": {
                  "type": "object",
                  "title": "Context"
               }
            },
            "type": "object",
            "required": [
               "loc",
               "msg",
               "type"
            ],
            "title": "ValidationError"
         }
      }
   }
}